Changes HTTPTest class name, test to cover HHTP String literal, change definitions HTTP.Header.key to use type alias HTTP.Headers
